api ReNaming & Controller 통합
